Made cleaner project folders layout.
