--- tests/test01_out.txt
@@ -1,6 +1,0 @@
-[0,0]: ab1AB
-[0,1]: cd2CD
-[0,2]: ef3EF
-[1,0]: gh4GH
-[1,1]: ij5IJ
-[1,2]: kl6KL
